add an ability to specify filetype
there is a problem when you try to open a file from url
in which case it doesn't have an extension.
roo gem has the ability to force specific spreadsheet type,
so i've added the :filetype option which can be used to set file type
manually.
also i've extracted explicit arguments to options hash,
tests are included