refactoring
I wanted to simplify things a bit, but it still should be backwards-compatible.
Also I had to remove some tests that contained knowledge about the internals.
It is possible now to use an instance of the `File` class with the `BatchFactory.from_file` method,
but at the moment there are no tests to prove it, going to write a couple soon